[ROS][refact] Synchronize with ros2

See merge request keu-computervision/slam!352

+++ slam_lib/src/SpinningSensorKeypointExtractor.cxx
@@ -472,7 +472,7 @@ void SpinningSensorKeypointExtractor::AddKptsUsingCriterion (Keypoint k,
                                                              double weightBasis)
 {
   // Loop over the scan lines
-  for (int scanlineIdx = 0; scanlineIdx < static_cast<int>(this->NbLaserRings); ++scanlineIdx)
+  for (unsigned int scanlineIdx = 0; scanlineIdx < this->NbLaserRings; ++scanlineIdx)
   {
     const PointCloud& scanlineCloud = *this->GetScanlineCloud(scanlineIdx);
     const int nPts = scanlineCloud.size();
@@ -552,7 +552,7 @@ void SpinningSensorKeypointExtractor::ComputeBlobs()
   std::mt19937 gen(2023); // Fix seed for deterministic processes
   std::uniform_real_distribution<> dis(0.0, 1.0);
 
-  for (unsigned int scanLine = 0; scanLine < static_cast<int>(this->NbLaserRings); ++scanLine)
+  for (unsigned int scanLine = 0; scanLine < this->NbLaserRings; ++scanLine)
   {
     const PointCloud& scanlineCloud = *this->GetScanlineCloud(scanLine);
     for (unsigned int index = 0; index < scanlineCloud.size(); ++index)
@@ -572,7 +572,7 @@ void SpinningSensorKeypointExtractor::EstimateAzimuthalResolution()
   // Compute horizontal angle values between successive points
   std::vector<float> angles;
   angles.reserve(this->Scan->size());
-  for (int scanLineIdx = 0; scanLineIdx < static_cast<int>(this->NbLaserRings); ++scanLineIdx)
+  for (unsigned int scanLineIdx = 0; scanLineIdx < this->NbLaserRings; ++scanLineIdx)
   {
     const auto& scanLineCloud = *this->GetScanlineCloud(scanLineIdx);
     for (unsigned int index = 1; index < scanLineCloud.size(); ++index)
